refactor(cli): separate large test suites
